Add Target API bindings for `python_app` and `jvm_app` (#9490)
We still need to implement hydration of the `bundles` field, but this PR means that we at least won't choke on encountering `python_app` and `jvm_app` in BUILD files.

There is some duplication between the two targets, but we don't try to generalize yet because it's not yet certain what support for `./pants bundle` will look like in V2 (e.g. if we even want to support it). We can always generalize later.

![python_app](https://user-images.githubusercontent.com/14852634/78733340-063c6580-78fa-11ea-919b-a49efeaedcb0.png)

![jvm_app](https://user-images.githubusercontent.com/14852634/78733360-13f1eb00-78fa-11ea-8bf4-5c6244492d68.png)

[ci skip-rust-tests]  # No Rust changes made.
[ci skip-jvm-tests]  # No JVM changes made.