Don't report diagnostics for cached targets
We tried reporting diagnostics for all target whether the cache was hit
or not, but this was observed not to scale in Twitter's infrastructure.

The amount of data written to the reporting server was simply too much.

To avoid this problem, handle the diagnostic counts as any other metric
(such as compile time) and only publish it to the metrics server on
recompiling a target.

The ownace is then on the queries to browse the history for relevant
context if necessary.

[ci skip-rust-tests]  # No Rust changes made.

[ci skip-jvm-tests]  # No JVM changes made.