Port AWS Lambda (V1 and V2) to use the Target API (#9491)
### V1 vs. V2 implementations

The implementations are different. V1 expects you to pass the field `binary` to specify a specific `python_binary`, whereas V2 is much more flexible in allowing you to specify `sources` and/or `dependencies` so that the `python_awslambda` target acts like a `python_binary` target. This results in much less boilerplate.

V1:

![Screen Shot 2020-04-07 at 11 02 31 PM](https://user-images.githubusercontent.com/14852634/78749850-defa8e00-7923-11ea-81d4-3b23cc5be291.png)

V2:

![Screen Shot 2020-04-07 at 11 03 08 PM](https://user-images.githubusercontent.com/14852634/78749881-f5084e80-7923-11ea-9669-c8616a5051b0.png)


In a followup, we will deprecate V1 AWS Lambda for the V2 implementation.


### Result: better error message on invalid target types

> ▶ ./v2 --no-print-exception-stacktrace awslambda src/python/pants/util:strutil
>
> ERROR: None of the provided targets work with the goal `awslambda`. This goal works with the following target types: ['python_awslambda'].

Note that this will only error if _no_ targets were valid. It will not error if _some_ targets are invalid and some are valid.

--

This also fixes bad `source` field entries in the root level BUILD.

[ci skip-rust-tests]
[ci skip-jvm-tests]