Port `./v2 test` to use the Target API (#9484)
### Result: better error messages for invalid target types with `--debug`

Before:

> ERROR: No valid test runner for src/python/pants/util:strutil.

After:

> ERROR: The `test` goal only works with the following target types: ['python_tests']
>
> You used src/python/pants/util:strutil with target type 'python_library'.

### Result: works with custom target types

So long as the target type has the field `PythonTestsSources` (or a subclass), it will work with `./v2 test`.

### Result: less hydration of sources

Previously, `HydratedTarget` would hydrate every target's sources no matter now. Now, we only hydrate sources for targets that we know are valid test targets (and for their transitive closure).

### Implementation: filtering out empty sources

When passed a target with empty sources, Pytest will run over all files in the transitive closure thanks to auto-discovery, which is wrong.

We need to filter out empty targets, which we continue to do in `test.py` to keep `pytest_runner.py` simpler. This requires hydrating the same sources in both `test.py` and `pytest_runner.py`, but the engine's caching should reduce the impact of this.

Likely, this mechanism should be factored up.