Fix test (#9493)
### Problem

The `test_unhashable_failure` test had a test case that depended on a specific line number in a Python exception coming from the Python standard library. This test worked on Python 3.6 but broke on Python 3.8 because that line number changed.

### Solution

Modify the test text to no longer depend on that specific line number. This makes the test pass regardless of the python 3 version.