Change precommit test (#9496)
### Problem

The code in the git hook for preparing a commit message that added the messages to skip rust CI/JVM CI/all CI wasn't checking to see if these sigils already existed in a commit message correctly. This resulted in commit messages automatically accumulating large numbers of these sigils if you repeatedly rebased a single commit.

### Solution

Fix the check to grep in the commit message for the exact text of the sigil, rather than trying to figure out from the second argument to the git hook whether this is an --amend commit or not. Also reduce some of the newlines that were printed to the commit message.