Fix 3.8 problem (#9502)
### Problem

In the `create` method of `core/test.py` there is a call to `issubclass` with a value that can in principle be a `typing.Union`. In Python 3.6 this works fine; however Python 3.8 changed the type of `typing.Union` in such a way as cause the `issubclass` call to throw an exception at runtime.

### Solution

Add a check to ensure that the value being checked by `issubclass` is an instance of `type` to avoid this error.