Create `@named_rule`s and add additional annotations to rules (#9453)
### Problem

We want to be able to create per-`@rule`-workunits that have additional metadata that comes from additional, optional annotations on `@rule`s.

### Solution

Create a new type of rule annotation called `@named_rule`, that will signal to the engine to create a workunit for the associated rule. `@named_rule`s take three kwargs, all of which are optional:

`name` - the canonical name of this rule. If not specified, this will default to the name of the python function.
`desc` - a human-readable description. If not specified, this will default to `None`. This will manifest in the workunit JSON as a string-valued key called `"desc"`.
`rule_type` - a member of a novel Enum that represents one of a small number of "types" of rules, meant to be used by clients of workunits to figure out what fields to expect and try to render. Defaults to `None` if not specified. This will manifest itself in the workunit JSON as a string-valued key called `"rule_type"`.