Add --pants-use-gitignore option (#9310)
### Problem

Cf. issue #5682, we would like pants to pay attention to a top-level .gitignore filee in a pants-controlled repository if the user passes in a `--pants-gitignore` flag.

### Solution

Pants already uses the Rust `ignore` crate to handle explicit ignore strings passed in the command line, so we can modify this code to also read `.gitignore` files when we try to do a file access, read the lines contained therein, and use that to decide whether a path should be ignored or not.