Distinguish pex sources from additional input files.
Currently all input files in PexRequest are placed under a
`source_files` subdir and passed to the pex binary using
the --sources-directory option.

However we may need to put files in the PEX building environment
that are not source files (e.g., a preamble file, or a prebuilt
requirements PEX).

This change:

- Renames PexRequest.input_files_digest to sources, for emphasis.
- Adds an additional_inputs field, and propagates those
  into the relevant digest.
- Makes corresponding changes in PexFromTargetsRequest.

[ci skip-rust-tests]
[ci skip-jvm-tests]