Don't show V1 bindings when exclusively using V2 (#9511)
### Problem

Now that we have bindings for every single core target and field, V2-only users have lots of irrelevant information presented in `target-types2`, such as this field being described on `python_library`:

```
    no_cache
        type: bool | None, default: False
        If True, don't store results for this target in the V1 cache.
```

### Solution

Just like we filter `./v2 goals` to only show V2 goals, this PR does the same thing for Fields and Targets.

If the user has `--v1` enabled, we will still show all V1 bindings. So, the default case does not change.

--

This solely impacts the output of `target-types2`. The V1 targets will still be registered and recognized when encountered in a BUILD file. This is important for mixed repos.

### Result

All the V2 targets can fit on one screen again!

![Screen Shot 2020-04-10 at 11 08 19 AM](https://user-images.githubusercontent.com/14852634/79012633-97d6ed80-7b1b-11ea-92ef-6bf7b6fd66b7.png)

[ci skip-rust-tests]
[ci skip-jvm-tests]