Do not mix platforms and interpreter constraints while building release pexes (#9515)
### Problem

The `release.sh -p` and `release.sh -q` modes build pexes from prebuilt wheels and from locally built wheels (respectively). But the mix of PEX flags that were in use in these two modes did not have the intended behavior for use with prebuilt wheels.

### Solution

When building from pre-built wheels, we're attempting to target a precise set of tags that match the wheels that were built. To do this, we should pass _only_ the `--platform` args that will match those wheels, but we were additionally passing `--interpreter-constraints`, which had the effect of additionally building for any other local pythons that we could find on the host machine.

When building wheels locally, we will have done so in the context of the python 3 binary located by the release script. In that case, we don't need intepreter constraints either, and can instead pass a particular python to PEX to target.

### Result

Consuming prebuilt wheels should succeed even in cases where a host machine has multiple python3 installs, some of which might have strange tag sets.

[ci skip-rust-tests]
[ci skip-jvm-tests]