Factor out a util function to parse a requirement constraint. (#9533)
This function supports our shorthand for omitting the interpreter
type (e.g., '>=3.7' means 'CPython>=3.7').

We needed this functionality in a second place, and so it made
sense to factor it out.

Instead of creating a new util.py, this change renames the existing
setup_py_util.py. Some of its existing functions were more general
purpose than the name implied anyway.  This change then adds the
new function to this repurposed util.py, and calls it from the
original location and from the new consumer (which happens to be
some other function in util.py.)

[ci skip-rust-tests]
[ci skip-jvm-tests]