Fix V2 binary still using the `current` platform when `platforms` is specified (#9563)
### Problem

When specifying `platforms` on a `python_binary`, you'd expect the built wheel to only have those specified platforms. But, it currently will always include the `current` platform.

This is because we are unconditionally passing the flag `--interpreter-constraint` to Pex, which forces `current` to be used, per https://github.com/pantsbuild/pex/issues/957. As explained there, Pex is behaving correctly; the issue is with Pants.

### Solution

When `platforms` are set, ignore interpreter constraints. 

Platforms already embed interpreter constraints in them, e.g. `linux-x86_64-cp-37-cp37m` saying that it needs compatibility with CPython 3.7. So, `platforms` is an override over interpreter constraints.

This change only applies when using the target with `./pants run` and `./pants binary`. Other goals like `./pants test` will ignore the `platforms` value and use interpreter constraints like normal.

[ci skip-rust-tests]
[ci skip-jvm-tests]