Unrevert black change and fix issue with it (#9545)
### Problem

#9536 broke a test that was making creating it's own build root and missed some config files.

It broke because on travis, we pass `PANTS_CONFIG_FILES=.../pants.travis-ci.toml` in the env. This test would change the buildroot, so that `pants.toml` et al were missing. This meant that black wasn't enabled because adding it happens in `pants.toml`.

### Solution

Convert the test to use `mock_buildroot`, which does the right thing.

### Result

The test passes.