Report workunits that have started but that are not yet complete (#9552)
### Problem

Clients of the StreamingWorkunits functionality should be able to get information about a workunit that has started but is not yet complete.

### Solution

Modify the workunit store to keep track of both when a workunit has been started, and when it has been marked as complete. This entailed a reasonably-substantial refactor of the workunit_store code, but also had the side effect of reducing the number of other crates that need to depend on `workunit_store`-internal data structures (with a little more refactoring around tests, we could probably make the `WorkUnit` type private). When the workunit store periodically calls a python FFI to update `StreamingWorkunit` clients, it now passes a list of started workunits as well as finished ones.