Use rust cpu_count crate to determine v2 UI swim lanes (#9574)
### Problem

The number of swim lanes displayed in the v2 UI is currently set to be the number of cpus on the local system, as determined by python code, and then passed through a couple of functions to the rust code that initializes the UI. It's not clear what the right number of swim lanes to have is; but either way that should be a concern of the rust UI code.

### Solution

Use the rust crate `cpu_count`, which pants rust code is already using, to determine the number of swim lanes when we initialize the v2 UI code; and remove the code paths that determine it in python. This moves the determination of the number of swim lanes closer to the UI code without changing the behavior of the v2 UI yet.