Add the ability to write junit test results file for pytest runs.
Problem

With the transition of pytest to the v2 engine, we lost the ability to write junit-xml results files when test runs.
This can't be solved by passing options thru directly to pytest since it expect a junit-xml result file name, which means we need a unique name for each target.

 Solution

Introduce the ability to calculate the junit-xml file name based on a user provided direcotry (option) and the test target name.

 Result

JUnit xml result files which can be processed by CI (CircleCI for example) so failing test can be seen w/o digging thru endless logs.