Add `DeduplicatedCollection` to newtype `FrozenOrderedSet` in rules (#9590)
## Problem

We already have `Collection` to create newtypes of single elements that work with the engine.

But, often, we need to deduplicate those elements for better cache hits, such as deduplicating the requirements we pass to Pex. Further, we often will want to sort the inputs for more cache hits.

## Solution

Add `engine.collection.DeduplicatedCollection` as a light wrapper around `FrozenOrderedSet`. Subclasses will inherit from this, just like `Collection`. 

It includes a mechanism for subclasses to signal that the inputs to their constructor should be sorted. This must be opted-into because it would be surprising to sort by default.

Note: we technically could just have classes directly subclass `FrozenOrderedSet`. But, this abstraction is more discoverable because it is colocated with `pants.engine.Collection` and it has some niceties like allowing sorting + giving a default value of an empty collection.

[ci skip-jvm-tests]
[ci skip-rust-tests]