Improve V2 UI messages by describing the scope of the work being done (#9588)
### Problem

Many of our Process descriptions do not fit on a single line because they list all the work they're doing, e.g. every target or every requirement. In https://github.com/pantsbuild/pants/pull/9534#issue-402926588, we decided the solution is not to hard-wrap multi-line entries.

We could make the descriptions more generic like `Resolving requirements`, or even `Resolving 10 requirements`, but we do want to preserve the specific work being done because it is possible to, for example, have two distinct PEXes with each only one requirement. It would be confusing to say "Resolving 1 requirement" both times because it looks like they're doing the same thing.

Rather, we want to preserve what specific work is being done, but de-prioritize it and ensure that we always explain the _scope_ of the work being done.

### Result

![lint example](https://user-images.githubusercontent.com/14852634/79775101-95794e00-82e8-11ea-869f-e94ba58a27a5.png)

![building pex with requirements](https://user-images.githubusercontent.com/14852634/79775332-f012aa00-82e8-11ea-8eda-c7ec7e7cedae.png)

[ci skip-rust-tests]
[ci skip-jvm-tests]