Factor up flexible handling of invalid target types (#9605)
## Problem

https://github.com/pantsbuild/pants/pull/9529 introduced the following flexible behavior for `binary` and `run`:

* Both `run` and `binary` will error if no valid targets were given.
* `run` will error if more than one _valid_ target was given. It doesn't care if there are invalid targets.
* `binary` will filter out all irrelevant targets.

The PR intentionally did not generalize the functionality because of the provisional status of the Configuration code. So, `test --debug` did not benefit and still would error if given more than one address, regardless of the address is irrelevant. `awslambda` also was duplicating the functionality of `binary` in a worse way.

## Solution

This PR generalizes the pattern and uses it with `awslambda`, `binary`, `run`, and `test`. We do not use it with `fmt`, `lint`, and `repl` because they all interact with the Target API differently.

## Result

`./v2 test --debug src/python/pants/util::` now works.

Invalid targets give nicer messages for `awslambda` and `test`:

```
▶ ./v2 --no-print-exception-stacktrace test 'src/python/pants/util/*util.py'


ERROR: The `test` goal only works with the following target types:
  * python_tests

You specified `src/python/pants/util/*util.py`, which only included the following target types:
  * files
  * python_library
```
[ci skip-rust-tests]
[ci skip-jvm-tests]