Temporarily disable `./pants backends` goal due to bug (#9629)
The goal works great when running in Pants' internal repo or repos with their own plugins, but doesn't work at all when Pants is installed as a wheel. This is because the implementation relies on `PathGlobs`, which only inspect the build root; it does not inspect installed wheels.

We need some way to inspect the Pants wheel. A stretch goal would inspect other installed Pants plugins.

For now, we skip it until coming up with a new strategy.

[ci skip-rust-tests]
[ci skip-jvm-tests]