Remoting: ensure oauth token file only has one line (#9630)
### Problem

The `remote_oauth_bearer_token_path` option is used in an Authorization header for remote execution requests, but is not validated to ensure it has only a single line.

### Solution

Validate the contents of the file loaded into the engine as the token, and remove any newlines on that single line.

### Result

I was able to trigger the validation code by invoking Pants with a remote_oauth_bearer_token_path file that had multiple lines.

Testing: I didn't see an existing unit test for this option. Would love to add one, just need suggestions as to where I should add it.