Fix selectors test for Python 3.8 (#9642)
### Problem

The `test_extract_constriants` test hard-coded the names of several ast nodes as reported by the Python standard library `ast` module, using the versions of these names for Python 3.6. Between Python 3.6 and Python 3.8, the names of these nodes changed slightly ("Num" and "Str" both became "Constant"). So the test was breaking on Python 3.8.

### Solution

Have the test compute the versions of these names according to the local interpreter version, ensuring that this test will pass for both versions of Python. This was tested locally under both Python 3.6 and Python 3.8.