Pass the entire Target to GenerateSourcesRequest
Stu and Benjy pointed out that it's common for a target to have fields on it that influence how the compiler runs. So, it will be common for the codegen rules to access the original target.

While we could expect rules to call `await Get[WrappedTarget](Address)` directly in the rules, this makes life simpler for them.
# Delete this line to force CI to run Clippy and the Rust tests.
[ci skip-rust-tests]
# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]