[jvm-platform] add strict as an attr to jvm platforms (#9287)
When a platform has been chosen as the default or otherwise, if there is an available JVM that's a higher version, that JVM may still be used. That can cause problems when you want to ensure that a particular platform always uses a specific JVM version.

Add strictness as a platform attribute, and apply it when looking up distributions.

Max versions for JVM lookups that use platforms can be bounded.