Handle source roots by stripping them
This contrasts with V1, which handles them via the env var proto_path. Stripping them is simpler because we already have several utils to do this. It is also more performant, as those utils have optimizations for calculating the source roots. Finally, this gives us parity with the V2 Python implementation.
# Delete this line to force CI to run Clippy and the Rust tests.
[ci skip-rust-tests]
# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]