Compute process execution times by walking the workunit graph (#9622)
### Problem

The V2 UI has been displaying an incorrect amount of time that a process has been running for some time now. The problem is that the time duration displayed in the UI is calculated from the real time at which a `Node` value was created. However, this is not necessarily the same as the time that a `Node` meaningfully starts to execute. Specifically in the case of `Process` nodes executed under the `BoundedCommandRunner`, a given `Node` might be blocked on executing until a previously-scheduled `Node` has finished, continuing to misleadingly count time that it is ostensibly executing even though it is not doing anything. 

### Solution

This commit moves the `heavy_hitters` logic that computes the longest-running node times from the `graph` crate to the `workunit_store` crate, and generally makes computing the amount of time a node has been running a responsibility of `workunit_store`. Similarly to the way the `heavy_hitters` worked previously, the new v2 UI logic creates a directed graph of workunits, then traverses to the sink nodes of the graph, and displays the top *k* of them, ordered by runningtime.

This, along with creating new workunits both within and around the `async_semaphor` lock in `BoundedCommandRunner`, ensures that, once a `Node` is actually running, the workunit used to render its duration will be a child of the workunit node representing the time it was waiting for the opportunity to execute. 

Note that it is still possible for a workunit representing a scheduled but not yet running `Process` to be a sink node of the workunit graph, and thus be displayed in the V2 UI. These workunits have the text "Bounded" in their title, which shows up in the UI (see the 2nd example below). It would be easy enough to filter these out so they are never displayed in the UI - we could for instance add a `display` boolean to `WorkunitMetadata`, set that to `false` on these bounded workunits, and filter for them in the `heavy_hitters` code. But I think it might be useful to end users to display these bounded nodes. Once they do in fact begin to run, the bounded workunit will have a child workunit that will be displayed instead, so the timing information will be accurate.

### Result

Here's what a hopefully-representative invocation of the `test` goal on some targets looks like with this new code:
https://asciinema.org/a/324524

And with the `process_execution_local_parallelism` option deliberately set to the low value of 2, in order to force `Process` nodes to wait on `BoundedCommandRunner`:
https://asciinema.org/a/324539