Always register `engine.target.rules()` via `engine_initializer.py` (#9661)
Certain rules get initialized in `engine_initializer.py` so that they are always registered, no matter what, such as in tests. We do this for the most fundamental rules, like the `fs.py` rules.

We originally didn't do this for `engine.target.rules()` because those rules were experimental, but they're now widely used. 

[ci skip-rust-tests]
[ci skip-jvm-tests]