Rename `Snapshot.directory_digest` to `Snapshot.digest` (#9658)
A `Digest` can be either "an opaque handle to a set of files" or it can refer to a string/bytes. To disambiguate, in the early days of V2, we used the variable name `directory_digest` to refer to when the digest is referring to the filesystem.

It turns out, we use `Digest` extremely frequently for the FS use. Because of this, we've over time shifted from `directory_digest` to the more ergonomic `digest`.

But, we're now inconsistent that we still sometimes use `directory_digest` and sometimes `digest`. This inconsistency results in wasted cognitive load.

Instead of fixing `digest` to say `directory_digest`, this PR converges around `digest`, given the overwhelming amount of use cases for the FS. If we ever need to use a `Digest` object to refer to a string/bytes, then we can rename that variable to `str_digest`.

This does not touch Rust, which does want the extra verbosity to disambiguate.

Closes https://github.com/pantsbuild/pants/issues/9129.

[ci skip-jvm-tests]