Change `GoalRuleTestBase.execute_rule` to return the captured stderr (#9664)
In an upcoming change, `repl` will write to stderr, rather than stdout. So, we need `GoalRuleTestBase` to allow returning both the stdout and stderr.

We use a dataclass to keep things organized.

Technically, `GoalRuleTestBase` was marked a public API, but this doesn't make much sense because the V2 API is nowhere near stable. We end up breaking V2 tests almost every single weekly release through changes to the engine. So, this removes that premature indicator that we weren't respecting in the first place.

[ci skip-rust-tests]
[ci skip-jvm-tests]