Build pantsbuild.pants as a manylinux2014 wheel (#9656)
Closes https://github.com/pantsbuild/pants/issues/7288. 

Centos6 is costly to support. It is hard to install Python 3.7 and Python 3.8 on it. Soon, we will no longer build a single `cp36-abi3` wheel, but will build a distinct wheel for each Python version. When making this change, we want a single CI job to build all the wheels in one go; so, we can no longer have Centos6 for building the Py36 wheel and Centos7 for building the Py37 and Py38 wheels.

See https://www.python.org/dev/peps/pep-0599/ for more information on manylinux2014.

[ci skip-rust-tests]
[ci skip-jvm-tests]