Remove noisy Pants output with V2 `binary`, `run`, and `repl` (#9667)
For `run`, there is no need for output. We simply display what the executable would normally display, and then return the same error code.

For `repl`, it should be obvious that we're starting the REPL and whether it was successful or not. The messages didn't add much.

For `binary`, the messages are useful. But, users need some way to disable them. And, we should write to stderr, rather than stdout. We use `logger.info` to achieve both goals.

[ci skip-rust-tests]
[ci skip-jvm-tests]