Improve `Snapshot::merge` errors. (#9685)
Previously just the basename of conflicting files was mentioned. Now
the full path of the conflicting file as well as the truncated contents
of each conflicting version of that file are included in the error
message. If Pants is run with debug logging enabled then full contents
of the conflicting files is included.