MyPy now checks partially_type_checked targets. (#9688)
The script was using bad `--tag` syntax. The `--tag` values are expected
to conform to `^[+-]?[^,]+(,[^,]+)*$`. In other words, a single leading
optional +/- operator and then a list of tag names to apply that single
operator to. As such, previously we were passing a `+` tag list
comprised of tag-literals `type_checked`, `+partially_type_checked` and
`-nolint`. Of these we only ever matched the `type-checked` tag literal.

Also fix up the type checking failures that this fix exposed.