Optimize `strip_source_roots.py` when there is only 1 source root (#9705)
As found in https://github.com/pantsbuild/pants/pull/9702#issuecomment-624740307, it is particularly expensive to run `SnapshotSubset`s (35.16 seconds for 10k files). There is also a small, but nontrivial cost to using `MergeDigests` (0.03 seconds for 10k files).

When we have only one single source root, there is no need for the extra work and extra complexity of trying to handle multiple source roots. Even if we optimize `SnapshotSubset`, we would still be doing unnecessary work.
 
Benchmark:

before | after | change
-- | -- | --
35.430 | 0.190 | -99%

(Running `strip_source_roots` on 10k files in a unit test.)

### Why still use `representative_path`?

We earlier added this as an optimization to only determine the source root for one single file in the snapshot, then use that result as the prefix for the whole snapshot. We use this when stripping a target's `sources`.

This optimization is still valuable. Benchmarking over 10k files:

* `representative_path`: 0.074
* no `representative_path`, after this PR: 0.190

After this PR, the `await Get[Digest](RemovePrefix)` ends up being identical to before. So, the cost comes from calling `itertools.groupby()` and calling `SourceRoots.find_by_path()` on every single file, instead of just one.

[ci skip-rust-tests]
[ci skip-jvm-tests]