Improve output of V2 `fmt` and `lint` goals (#9710)
### Problem

It's confusing which linters and formatters have run. For example, docformatter will simply put the file name with no context when there is a lint failure, like this:

```bash
$ ./v2 lint src/python::

src/python/pants/util/strutil.py
```

Goals of the design:
1. Disambiguate which linters were run.
2. Preserve the original output.
3. Be minimally intrusive to the user.

### Result

![fmt example](https://user-images.githubusercontent.com/14852634/81235842-c035f880-8fb0-11ea-92c5-6e2952b17f37.png)

![lint example](https://user-images.githubusercontent.com/14852634/81236042-30447e80-8fb1-11ea-92f8-ec4e30b86ff6.png)

We always write to stderr for consistency and to follow the Unix philosophy. Because the output mangles multiple tools into one stream, the output cannot be safely piped to other tools. (Instead, we should add support for writing tool output to specific files, like we allow with Pytest.)

[ci skip-rust-tests]
[ci skip-jvm-tests]