Fix up dependency features
As per https://github.com/rust-lang/cargo/issues/4463 if this package
happens to be built via the `engine/Cargo.toml` it will succeed (because
the `derive` feature will be enabled), if built separately it will fail.

This enables the feature explicitly in the crate's `Cargo.toml`,
reflecting that the code accesses the re-exported version of the trait.
[ci skip-jvm-tests]