Reorganize some V2 goals into `pants.backend.pants_info` and `pants.backend.project_info` (#9708)
We want to distinguish between rules that a) give information on Pants internals and b) give information on the project internals.

With both, we want the backends to be activated by default, but to also be able to be opted out of. (Unlike `pants.core`, which must never be opted out of to avoid breaking Pants.)

So, we reorganize `core/project_info/` and `backend/project_info/` into top-level `backend/pants_info/` and `backend/project_info/` folders.