Introduce resolve proof of concept.
This adds a goal only used for experimentation so far. It uses the new Pex
download APIs via the resolver binary to resolve distributions without building
them and it also maintains caches without introducing any new intrinsics by
recursively resolving targets and using the result of resolves to form the
cache for the dependent resolves.
# Delete this line to force CI to run Clippy and the Rust tests.
[ci skip-rust-tests]
# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]