Restructure v2 UI (#9672)
### Problem

The V2 UI currently takes over the entire terminal, which is problematic for scrollback and for redirecting pants output, and is just generally fragile bespoke code. We'd like to be able to display a dynamically-updating UI while pants is running without breaking these things.

### Solution

Replace the existing UI with a new type `ConsoleUI` that displays a UI similar to the existing one - that is, a series of swim lanes with log messages written above them. This new type uses the [Indicatif](https://github.com/mitsuhiko/indicatif) library for rendering the dynamically-updating swim lanes.

Additionally, the lightning bolt icons have been replaced by an Indicatif-provided dynamically-spinning widget, which looks a little more professional and helps indicate that pants is currently doing something.

### Result

The following asciinema terminal recordings show:

Ordinary pants usage (https://asciinema.org/a/329317)
Piping the stdout output of pants to a file: (https://asciinema.org/a/329321)
And hitting Ctrl-C to halt pants during the middle of a run (https://asciinema.org/a/329328)

Thanks to @stuhood for help with integrating the Indicatif display-updating thread with the rest of pants.