Fix write_stderr in native.py (#9753)
### Problem

The `write_stderr` function in `native.py` was mistakenly calling the Rust FFI function to write to stdout rather than to write to stderr. This was causing problems with redrawing swim lanes in the new terminal UI, and is in general incorrect.

### Solution

Have the `native.py` function invoke the correct Rust FFI function.