Disable capturing core dumps in travis. (#9769)
### Problem

A while back we started capturing core dumps "globally" in travis. But in practice we have never consumed them, and I'm fairly certain that they are causing the OSX shards that test sending `SIGABRT` (which, if core dumps are enabled, will trigger a core dump) to `pantsd` to:
1. be racey, because while the core is dumping, the process is non-responsive and can't be killed, leading to errors like:
```FAILURE: failure while terminating pantsd: failed to kill pid 28775 with signals (<Signals.SIGTERM: 15>, <Signals.SIGKILL: 9>)```
2. run out of disk space: we've seen mysterious "out of disk" errors on the OSX shards... and core dumps are large.

### Solution

Disable core dumps everywhere. If we end up needing them in the future, we can enable them on a case-by-case basis.

### Result

Fixes #8127.

[ci skip-rust-tests]
[ci skip-jvm-tests]